spa1 to spa2 veh transfer

+++ SPAII Setting/frmSPA.vb
@@ -37,6 +37,12 @@
             cbOnlineMap.Checked = True
         End Try
 
+        Try
+            cbMission.Checked = CBool(ReadIniValue(config, "SETTING", "HideBlipsOnMission"))
+        Catch ex As Exception
+            cbMission.Checked = True
+        End Try
+
         txtSPA1.Text = IO.Path.GetFullPath("..\SinglePlayerApartment\Garage")
         txtSPA2.Text = $"{My.Application.Info.DirectoryPath}\Garages"
         cmbSPA1.Items.AddRange(SPA1Property.Values.ToArray)
@@ -62,6 +68,7 @@
 
         WriteIniValue(config, "SOUND", "Volume", tbVolume.Value)
         WriteIniValue(config, "SETTING", "OnlineMap", cbOnlineMap.Checked)
+        WriteIniValue(config, "SETTING", "HideBlipsOnMission", cbMission.Checked)
 
         MsgBox("Settings saved.", MsgBoxStyle.Information, "SPA II")
     End Sub
@@ -85,6 +92,7 @@
             For Each file As String In IO.Directory.GetFiles(selectedItem, "*.cfg")
                 Dim item As New ListViewItem(ReadCfgValue(Of String)("VehicleName", file))
                 With item
+                    .SubItems.Add(ReadCfgValue(Of String)("PlateNumber", file))
                     .SubItems.Add(IO.Path.GetFileName(file))
                     .Tag = file
                 End With
@@ -101,8 +109,9 @@
             Dim selectedItem As String = $"{txtSPA2.Text}\{SPA2Property.Where(Function(x) x.Value(0) = cmbSPA2.SelectedItem).FirstOrDefault.Key}"
             For Each file As String In IO.Directory.GetFiles(selectedItem, "*.xml")
                 Dim vd As VehicleData = New VehicleData(file).Instance
-                Dim item As New ListViewItem(vd.Name)
+                Dim item As New ListViewItem($"{vd.Make} {vd.Name}")
                 With item
+                    .SubItems.Add(vd.PlateNumber)
                     .SubItems.Add(IO.Path.GetFileName(file))
                     .Tag = file
                 End With
@@ -114,28 +123,48 @@
     End Sub
 
     Private Sub btnTransfer_Click(sender As Object, e As EventArgs) Handles btnTransfer.Click
-        Dim spa2 = SPA2Property.Where(Function(x) x.Value(0) = cmbSPA2.SelectedItem).FirstOrDefault
-        Dim spa2Path = spa2.Key
-        Dim spa2Garage = spa2.Value
-        Dim spa2GaragePath = spa2Garage(0)
-        Dim spa2GarageSize As Integer = spa2Garage(1)
-        Dim spa2PropertyID As Integer = spa2Garage(2)
-        Dim selectedCount As Integer = lvSPA1.SelectedItems.Count
-        Dim spa2GarageVehCount As Integer = lvSPA2.Items.Count
-        Dim spaceLeft As Integer = spa2GarageSize - (spa2GarageVehCount - selectedCount)
+        Try
+            If lvSPA1.SelectedItems.Count = 0 Then
+                MsgBox("Please select at least 1 vehicle to transfer.", MsgBoxStyle.Exclamation, "ERR")
+            ElseIf cmbSPA1.SelectedIndex = -1 OrElse cmbSPA2.SelectedIndex = -1 Then
+                MsgBox("Please select garage.", MsgBoxStyle.Exclamation, "ERR")
+            Else
+                Dim spa2 = SPA2Property.Where(Function(x) x.Value(0) = cmbSPA2.SelectedItem).FirstOrDefault
+                Dim spa2Path = spa2.Key
+                Dim spa2Garage = spa2.Value
+                Dim spa2PropName = spa2Garage(0)
+                Dim spa2GarageSize As Integer = spa2Garage(1)
+                Dim spa2PropertyID As Integer = spa2Garage(2)
+                Dim selectedCount As Integer = lvSPA1.SelectedItems.Count
+                Dim spa2GarageVehCount As Integer = lvSPA2.Items.Count
+                Dim spaceLeft As Integer = spa2GarageSize - (spa2GarageVehCount - selectedCount)
 
-        If spaceLeft >= selectedCount Then
-            For Each item As ListViewItem In lvSPA1.SelectedItems
-                Dim spa1File As String = item.Tag
-                Dim spa2Veh As New VehicleData(spa2GaragePath, spa1File, GetAvailableIndex(Vehicles(spa2Path), spa2GarageSize), spa2PropertyID)
-                spa2Veh.Save()
-                If IO.File.Exists(spa1File) Then IO.File.Delete(spa1File)
-            Next
+                If spaceLeft >= selectedCount Then
+                    For Each item As ListViewItem In lvSPA1.SelectedItems
+                        Dim spa1File As String = item.Tag
+                        Dim spa2Veh As New VehicleData(spa2Path, spa1File, GetAvailableIndex(Vehicles($"{txtSPA2.Text}\{spa2Path}"), spa2GarageSize), spa2PropertyID)
+                        spa2Veh.Save()
+                        If IO.File.Exists(spa1File) Then IO.File.Delete(spa1File)
+                    Next
 
-            SPA1Load()
-            SPA2Load()
-        Else
-            MsgBox("Your garage is full, cannot perform this action", MsgBoxStyle.Exclamation, "ERR_GRG_FULL")
-        End If
+                    SPA1Load()
+                    SPA2Load()
+                Else
+                    MsgBox("Your garage is full, cannot perform this action", MsgBoxStyle.Exclamation, "ERR_GRG_FULL")
+                End If
+            End If
+        Catch ex As Exception
+            MsgBox($"{ex.Message} {ex.StackTrace}", MsgBoxStyle.Critical, "ERR")
+        End Try
+    End Sub
+
+    Private Sub btnBrowse1_Click(sender As Object, e As EventArgs) Handles btnBrowse1.Click
+        Dim fsd = New FolderSelectDialog() With {.Title = "Select Garage Directory...", .InitialDirectory = txtSPA1.Text}
+        If fsd.ShowDialog(IntPtr.Zero) Then txtSPA1.Text = fsd.FileName
+    End Sub
+
+    Private Sub btnBrowse2_Click(sender As Object, e As EventArgs) Handles btnBrowse2.Click
+        Dim fsd = New FolderSelectDialog() With {.Title = "Select Garage Directory...", .InitialDirectory = txtSPA2.Text}
+        If fsd.ShowDialog(IntPtr.Zero) Then txtSPA2.Text = fsd.FileName
     End Sub
 End Class
